refactor(tui): extract app formatters constants types and keys
