+++ tui/src/app/constants.ts
@@ -1,0 +1,7 @@
+export const SEARCH_DEBOUNCE_MS = 500
+export const SEARCH_RESULT_LIMIT = 25
+export const LIST_SCROLLBAR_WIDTH = 2
+export const STATUS_BAR_PADDING_X = 1
+export const SPLIT_LAYOUT_MIN_WIDTH = 100
+export const SPINNER_FRAME_MS = 90
+export const BRAILLE_SPINNER_FRAMES = ['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '⠏'] as const
